Fixing plan and income services

+++ src/main/java/disenodesistemas/backendfunerariaapp/service/impl/IncomeServiceImpl.java
@@ -5,11 +5,13 @@ import disenodesistemas.backendfunerariaapp.dto.request.IncomeDetailRequestDto;
 import disenodesistemas.backendfunerariaapp.dto.response.IncomeResponseDto;
 import disenodesistemas.backendfunerariaapp.entities.IncomeDetailEntity;
 import disenodesistemas.backendfunerariaapp.entities.IncomeEntity;
+import disenodesistemas.backendfunerariaapp.entities.ItemEntity;
 import disenodesistemas.backendfunerariaapp.entities.ReceiptTypeEntity;
 import disenodesistemas.backendfunerariaapp.entities.SupplierEntity;
 import disenodesistemas.backendfunerariaapp.entities.UserEntity;
 import disenodesistemas.backendfunerariaapp.exceptions.AppException;
 import disenodesistemas.backendfunerariaapp.repository.IncomeRepository;
+import disenodesistemas.backendfunerariaapp.repository.ItemRepository;
 import disenodesistemas.backendfunerariaapp.service.Interface.IncomeService;
 import disenodesistemas.backendfunerariaapp.service.Interface.SupplierService;
 import disenodesistemas.backendfunerariaapp.service.Interface.UserService;
@@ -33,8 +35,11 @@ import org.springframework.transaction.annotation.Transactional;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import java.util.stream.Collectors;
 
+import static java.util.Objects.nonNull;
 import static org.apache.commons.collections.CollectionUtils.isEmpty;
 
 
@@ -44,6 +49,7 @@ import static org.apache.commons.collections.CollectionUtils.isEmpty;
 public class IncomeServiceImpl implements IncomeService {
 
     private final IncomeRepository incomeRepository;
+    private final ItemRepository itemRepository;
     private final UserService userService;
     private final SupplierService supplierService;
     private final ProjectionFactory projectionFactory;
@@ -61,7 +67,10 @@ public class IncomeServiceImpl implements IncomeService {
                 .build();
         checkExistsByByReceiptNumber(incomeEntity, incomeRequestDto);
         if (!isEmpty(incomeRequestDto.getIncomeDetails())) {
-            incomeEntity.setIncomeDetails(modelMapper.map(incomeRequestDto.getIncomeDetails(), new TypeToken<List<IncomeDetailEntity>>() {}.getType()));
+            incomeRepository.save(incomeEntity);
+            deleteIncomeWhereIsDeletedTrue(incomeEntity);
+            incomeEntity.setIncomeDetails(getIncomeDetailsEntities(incomeRequestDto.getIncomeDetails()));
+            setItemPrices(incomeEntity.getIncomeDetails());
         }
         totalAmountCalculator(incomeEntity, incomeRequestDto);
         return projectionFactory.createProjection(IncomeResponseDto.class, incomeRepository.save(incomeEntity));
@@ -69,8 +78,9 @@ public class IncomeServiceImpl implements IncomeService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<IncomeResponseDto> getAllIncomes() {
-        return incomeRepository.findAllByOrderByIdDesc();
+    public List<IncomeResponseDto> getAllIncomes(final Boolean deleted) {
+        return Boolean.TRUE.equals(deleted) ? incomeRepository.findAllByOrderByIdDesc()
+                : incomeRepository.findByDeletedFalseOrderByIdDesc();
     }
 
     @Override
@@ -89,6 +99,8 @@ public class IncomeServiceImpl implements IncomeService {
             final List<IncomeDetailEntity> incomeDetailEntities = getDeletedIncomeDetails(incomeEntity, incomeRequestDto);
             incomeDetailEntities.forEach(incomeEntity::removeIncomeDetail);
             incomeEntity.setIncomeDetails(modelMapper.map(incomeRequestDto.getIncomeDetails(), new TypeToken<List<IncomeDetailEntity>>() {}.getType()));
+            deleteIncomeWhereIsDeletedTrue(incomeEntity);
+            setItemPrices(incomeEntity.getIncomeDetails());
         }
         return projectionFactory.createProjection(IncomeResponseDto.class, incomeRepository.save(incomeEntity));
     }
@@ -97,11 +109,15 @@ public class IncomeServiceImpl implements IncomeService {
     @Transactional
     public void deleteIncome(final Long receiptNumber) {
         val incomeEntity = findEntityByReceiptNumber(receiptNumber);
+        if (incomeEntity.isDeleted()) {
+            incomeRepository.delete(incomeEntity);
+            throw new AppException("income.error.already.deleted", HttpStatus.CONFLICT);
+        }
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         final String email = authentication.getName();
         incomeEntity.setLastModifiedBy(userService.getUserByEmail(email));
         incomeRepository.save(incomeEntity);
-        incomeRepository.delete(incomeEntity);
+        incomeRepository.deleteByReceiptNumber(incomeEntity.getReceiptNumber());
     }
 
     @Override
@@ -142,13 +158,46 @@ public class IncomeServiceImpl implements IncomeService {
     }
 
     private void checkExistsByByReceiptNumber(final IncomeEntity incomeEntity, final IncomeRequestDto incomeDto) {
-        if(incomeRepository.existsByReceiptNumber(incomeDto.getReceiptNumber()))
+        final boolean isDeleted = incomeRepository.findByReceiptNumber(incomeDto.getReceiptNumber())
+                .map(IncomeEntity::isDeleted).orElse(true);
+        if(!isDeleted)
             throw new AppException("income.error.receiptNumber.already.registered", HttpStatus.CONFLICT);
         incomeEntity.setReceiptNumber(incomeDto.getReceiptNumber());
     }
 
     private IncomeEntity findEntityByReceiptNumber(final Long receiptNumber) {
         return incomeRepository.findByReceiptNumber(receiptNumber).orElseThrow(() -> new AppException("income.error.not.found", HttpStatus.NOT_FOUND));
+    }
+
+    private List<IncomeDetailEntity> getIncomeDetailsEntities(final List<IncomeDetailRequestDto> incomeDetailsRequestDto) {
+        final List<ItemEntity> itemEntities = (List<ItemEntity>) itemRepository.findAll();
+        return incomeDetailsRequestDto.stream().map(
+                income -> {
+                    final ItemEntity itemEntity = itemEntities.stream()
+                            .filter(item -> income.getItem().getCode().equals(item.getCode()))
+                            .findFirst().orElse(null);
+                    final IncomeDetailEntity incomeToReturn = modelMapper.map(income, IncomeDetailEntity.class);
+                    incomeToReturn.setItem(itemEntity);
+                    return incomeToReturn;
+                }
+        ).collect(Collectors.toUnmodifiableList());
+    }
+
+    private void setItemPrices(final List<IncomeDetailEntity> incomeDetails) {
+        final List<ItemEntity> itemsToUpdatePrice = incomeDetails.stream().filter(Objects::nonNull)
+                .map(incomeDetail -> {
+                    incomeDetail.getItem().setPrice(incomeDetail.getSalePrice());
+                    return incomeDetail.getItem();
+                }).collect(Collectors.toUnmodifiableList());
+        itemRepository.saveAll(itemsToUpdatePrice);
+    }
+
+    private void deleteIncomeWhereIsDeletedTrue(final IncomeEntity incomeEntity) {
+        final IncomeEntity incomeEntityToDelete = incomeRepository.findByReceiptNumberOrderByIdDesc(incomeEntity.getReceiptNumber())
+                .stream()
+                .filter(IncomeEntity::isDeleted)
+                .findFirst().orElse(null);
+        if (nonNull(incomeEntity)) incomeRepository.delete(incomeEntityToDelete);
     }
 
 }
